+++ planck-c/main.c
@@ -298,7 +298,38 @@ bool should_ignore_arg(const char *opt) {
             last_c == 'L');
 }
 
+void control_FTL_JIT() {
+
+    /* Recent versions of JavaScriptCore are crashing in FTL JIT. At least try
+     * to avoid this on macOS for now: If on macOS 10.13.4 (or a later 10.13 version)
+     * and JSC_useFTLJIT env var not set, disable FTL JIT */
+
+    if (getenv("JSC_useFTLJIT") == NULL) {
+
+        FILE* fp = fopen("/System/Library/CoreServices/SystemVersion.plist", "r");
+        char* version_info = read_all(fp);
+        fclose(fp);
+
+        char* version_str = strstr(version_info, "<string>10.13");
+        if (version_str) {
+            version_str += 14;
+
+            int version = atoi(version_str);
+
+            if (version >= 4) {
+                putenv("JSC_useFTLJIT=false");
+            }
+        }
+
+        free(version_info);
+    }
+}
+
 int main(int argc, char **argv) {
+
+#ifdef __APPLE__
+    control_FTL_JIT();
+#endif
 
     ignore_sigpipe();